pyproject.toml
[project]
name = "wiedunflow"
version = "0.6.0"
description = "Generate interactive, self-contained HTML tutorials from local Git repositories"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Michał Kamiński", email = "michal.s.kaminski@gmail.com" }]
keywords = ["tutorial", "documentation", "llm", "code-analysis", "ast", "rag", "anthropic", "openai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Documentation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.1.7",
    "pydantic>=2.7",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1",
    "pygments>=2.17",
    "markupsafe>=2.1",
    "tree-sitter>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "jedi>=0.19.2",
    "networkx>=3.3",
    "pathspec>=0.12.1",
    "structlog>=24.1",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "rank-bm25>=0.2.2",
    "tenacity>=9.0.0",
    "platformdirs>=4.3.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "rich>=13.7",
    "mistune>=3.2.0",
    "questionary>=2.1.1,<3.0",
    "httpx>=0.27",
]

[project.urls]
Homepage = "https://github.com/numikel/wiedunflow"
Repository = "https://github.com/numikel/wiedunflow"
Changelog = "https://github.com/numikel/wiedunflow/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/numikel/wiedunflow/issues"

[project.scripts]
wiedun-flow = "wiedunflow.cli.main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-playwright>=0.5",
    "pytest-socket>=0.7",
    "syrupy>=4.6",
    "mypy>=1.10",
    "ruff>=0.5.0",
    "pre-commit>=3.8",
    "pyyaml>=6.0",
    "types-pyyaml>=6.0",
    "pillow>=10.0",
    "pip-audit>=2.7",
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wiedunflow"]

[tool.hatch.build.targets.wheel.force-include]
"src/wiedunflow/renderer/fonts" = "wiedunflow/renderer/fonts"
"src/wiedunflow/renderer/templates" = "wiedunflow/renderer/templates"

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]
extend-exclude = [
    "tests/eval/corpus/repos",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM", "RUF", "PL", "TCH"]
ignore = [
    "PLR0913",
    "E501",
    "TC001",
    "TC002",
    "TC003",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["PLR2004", "S101", "PLR0915", "PLC0415", "PLW0108"]
"__init__.py" = ["F401"]
# menu.py uses local imports inside helpers to avoid an import cycle with main.py
# (main.main() loads menu lazily; menu._launch_pipeline() loads main lazily back).
# Sub-wizards have many branch / return paths because each prompt boundary can abort.
"src/wiedunflow/cli/menu.py" = ["PLC0415", "PLR0911", "PLR0912", "PLR0915", "SIM102"]
# scripts/* are utility one-shots, not production library code.
"scripts/**/*.py" = ["PLR2004"]
# config._normalize_target_audience has explicit per-prefix branches for the
# fuzzy-mapping shim; flattening it into a dict harms readability.
"src/wiedunflow/cli/config.py" = ["PLR0911"]
# main.main() lazy-imports the menu module to avoid an import cycle.
"src/wiedunflow/cli/main.py" = ["PLC0415"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_decorators = false

[[tool.mypy.overrides]]
module = ["jedi.*", "jedi"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["networkx.*", "networkx"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["rank_bm25"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-m", "not eval and not eval_robustness",
    "--ignore=tests/eval/corpus/repos",
]
testpaths = ["tests"]
norecursedirs = ["tests/eval/corpus/repos", ".git", ".venv"]
markers = [
    "eval: end-to-end tests that call real LLM APIs (require ANTHROPIC_API_KEY, excluded from default run)",
    "eval_robustness: analysis-only robustness tests on 5 pinned OSS repos with FakeLLMProvider (opt-in, log-only)",
    "integration: walking skeleton e2e tests",
    "playwright: headless browser tests (require playwright install)",
    "netns: Linux-only zero-telemetry tests using unshare -n (US-011)",
]
